+++ src/components/layout/Menu.vue
@@ -107,12 +107,6 @@
                                 icon: 'fas fa-search'
                             },
                             {
-                                href: '/pms/libmapping',
-                                hidden: wtutils.hideMenu('pmsLibMapping'),
-                                title: this.$t("Common.Menu.Sidebar.PMS.LibMapping"),
-                                icon: 'fas fa-hdd'
-                            },
-                            {
                                 href: '/pms/settings',
                                 hidden: wtutils.hideMenu('pmsSettings'),
                                 title: this.$t("Common.Menu.Sidebar.PMS.Settings"),
@@ -139,7 +133,7 @@
                         icon: 'fas fa-globe'
                     },
                     {
-                        href: '/settings',
+                        href: '/settings/settings',
                         title: this.$t("Common.Menu.Sidebar.Settings.NavTitle"),
                         hidden: wtutils.hideMenu('Settings'),
                         icon: 'fas fa-cog'
