+++ src/components/modules/ExportTools/Settings/settings.vue
@@ -4,18 +4,11 @@
     <p id="demo"></p>
     <h1>{{ $t("Modules.ET.Settings.Name") }}</h1>
     <p>{{ $t("Modules.ET.Settings.Description") }}</p>
-    <b-input-group id="ArraySepGrp" :prepend="$t('Modules.ET.Settings.ArraySep')" class="mt-3">
-        <b-form-input id="ArraySep" name="ArraySep" type="text" class="form-control" v-model="ArraySep" :disabled=false :maxlength=1 @change="setArraySep()"></b-form-input>
-    </b-input-group>
-    <b-input-group id="setTextQualifierCSVGrp" :prepend="$t('Modules.ET.Settings.QualifierCSV')" class="mt-3">
-        <b-form-input id="TextQualifierCSV" name="TextQualifierCSV" type="text" class="form-control" v-model="TextQualifierCSV" :disabled=false :maxlength=1 @change="setTextQualifierCSV()"></b-form-input>
-    </b-input-group>
-    <b-input-group id="NotAvailIndicatorGrp" :prepend="$t('Modules.ET.Settings.NotAvailIndicator')" class="mt-3">
-        <b-form-input id="NotAvailIndicator" name="NotAvailIndicator" type="text" class="form-control" v-model="NotAvailIndicator" :disabled=false @change="setNotAvailIndicator()"></b-form-input>
-    </b-input-group>
-    <b-input-group id="ColumnSepGrp" :prepend="$t('Modules.ET.Settings.ColumnSep')" class="mt-3">
-        <b-form-input id="ColumnSep" name="ColumnSep" type="text" class="form-control" v-model="ColumnSep" :disabled=false :maxlength=1 @change="setColumnSep"></b-form-input>
-    </b-input-group>
+    <b-link id="general" to="/settings/export">{{ $t("Modules.ET.Settings.Note") }} </b-link>
+    <br>
+
+
+
     <b-input-group id="PosterGrp" :prepend="$t('Modules.ET.Settings.Posters_Dimensions')" class="mt-3">
         <b-tooltip target="PosterGrp" triggers="hover">
               {{ $t('Modules.ET.Settings.Posters_Dimensions_TT') }}
@@ -27,18 +20,6 @@
               {{ $t('Modules.ET.Settings.Art_Dimensions_TT') }}
         </b-tooltip>
         <b-form-input id="ArtDim" name="ArtDim" type="text" class="form-control" v-model="ArtDim" :disabled=false @change="setArt_Dimensions()"></b-form-input>
-    </b-input-group>
-    <b-input-group id="ChReturn" :prepend="$t('Modules.ET.Settings.ChReturn')" class="mt-3">
-        <b-tooltip target="ChReturn" triggers="hover">
-              {{ $t('Modules.ET.Settings.ChReturn_TT') }}
-        </b-tooltip>
-        <b-form-input id="ChReturn" name="ChReturn" type="text" class="form-control" v-model="ChReturn" :disabled=false @change="setChReturn()"></b-form-input>
-    </b-input-group>
-    <b-input-group id="ChNewLine" :prepend="$t('Modules.ET.Settings.ChNewLine')" class="mt-3">
-        <b-tooltip target="ChNewLine" triggers="hover">
-              {{ $t('Modules.ET.Settings.ChNewLine_TT') }}
-        </b-tooltip>
-        <b-form-input id="ChNewLine" name="ChNewLine" type="text" class="form-control" v-model="ChNewLine" :disabled=false @change="setChNewLine()"></b-form-input>
     </b-input-group>
     <b-form-group id="b-form-group">
       <b-form-checkbox-group
@@ -73,23 +54,11 @@
     export default {
         created() {
             this.getDefaults();
-            if (wtconfig.get('ET.ColumnSep') == '\t')
-            {
-                this.ColumnSep = '{TAB}';
-            }
-            else
-            {
-                this.ColumnSep = wtconfig.get('ET.ColumnSep');
-            }
             // alert('For this version, export to XLSX is currently disabled');
             this.showAlert();
         },
         data() {
             return {
-                ArraySep: wtconfig.get('ET.ArraySep'),
-                TextQualifierCSV: wtconfig.get('ET.TextQualifierCSV', '"'),
-                NotAvailIndicator: wtconfig.get('ET.NotAvail', 'N/A'),
-                ColumnSep: '',
                 PosterDim: wtconfig.get('ET.Posters_Dimensions', '75*75'),
                 ArtDim: wtconfig.get('ET.Art_Dimensions', '75*75'),
                 cbSelected: [],
@@ -101,8 +70,6 @@
                     { text: i18n.t('Modules.ET.Settings.suggestedFileNoExtra'), value: 'suggestedFileNoExtra' },
                     { text: i18n.t('Modules.ET.Settings.suggestedUseOrigenTitle'), value: 'suggestedUseOrigenTitle' }
                 ],
-                ChReturn: wtconfig.get('ET.ChReturn', '<RETURN>'),
-                ChNewLine: wtconfig.get('ET.ChNewLine', '<NEWLINE>'),
                 SelectedMoviesIDOptions: ['imdb', 'tmdb'],
                 SelectedMoviesID: '',
                 SelectedShowsIDOptions: ['tmdb', 'tvdb'],
@@ -130,45 +97,6 @@
                 for( var cbItem of ["ExpCSV","ExpXLSX","OrgTitleNull", "SortTitleNull", "AutoXLSCol", "AutoXLSRow", "suggestedFileNoExtra", "suggestedUseOrigenTitle"]){
                     wtconfig.set("ET." + cbItem, (this.cbSelected.includes(cbItem)))
                 }
-            },
-            setColumnSep(val){
-                if (val.length > 1)
-                {
-                    this.$bvToast.toast(this.$t("Modules.ET.ErrorBadSep"), {
-                        title: this.$t("Modules.ET.ErrorBadSepTitle"),
-                        autoHideDelay: 3000,
-                        solid: true,
-                        variant: 'primary',
-                        toaster: 'b-toaster-bottom-center'
-                    });
-                }
-                else
-                {
-                    if (val == '9')
-                    {
-                        wtconfig.set('ET.ColumnSep', '\t')
-                        this.ColumnSep = '{TAB}'
-                    }
-                    else
-                    {
-                        wtconfig.set('ET.ColumnSep', this.ColumnSep)
-                    }
-                }
-            },
-            setArraySep: function(){
-                wtconfig.set('ET.ArraySep', this.ArraySep)
-            },
-            setChReturn: function(){
-                wtconfig.set('ET.ChReturn', this.ChReturn)
-            },
-            setChNewLine: function(){
-                wtconfig.set('ET.ChNewLine', this.ChNewLine)
-            },
-            setTextQualifierCSV: function(){
-                wtconfig.set('ET.TextQualifierCSV', this.TextQualifierCSV)
-            },
-            setNotAvailIndicator: function(){
-                wtconfig.set('ET.NotAvail', this.NotAvailIndicator)
             },
             setPosters_Dimensions: function(){
                 wtconfig.set('ET.Posters_Dimensions', this.PosterDim);
