+++ src/components/modules/General/wtutils.js
@@ -262,6 +262,9 @@ const wtutils = new class WTUtils {
         // Update config file with defaults if missing
         log.verbose('[wtutils.js] (UpdateConfigFile) Updating config file');
         // Hide/Show Menu Set to false if enabled for all
+        if ( wtconfig.get('Menu.pmsFindMedia', 'N/A') == 'N/A' ){
+            wtconfig.set('Menu.pmsFindMedia', false)
+        }
         if ( wtconfig.get('Menu.plextv', 'N/A') == 'N/A' ){
             wtconfig.set('Menu.plextv', false)
         }
