+++ src/router/index.js
@@ -10,11 +10,14 @@ import PMS from '../components/modules/PMS/PMS';
 import PMSSettings from '../components/modules/PMS/Settings/settings';
 import Butler from '../components/modules/PMS/Butler/butler';
 import Language from '../components/modules/Main/Language.vue';
-import GlobalSettings from '../components/modules/Main/GlobalSettings';
+import SettingsGlobal from '../components/modules/Main/Settings/Settings';
 import DVR from '../components/modules/PMS/DVR/DVR';
 import FindMedia from '../components/modules/PMS/FindMedia/FindMedia';
 import FindMediaSettings from '../components/modules/PMS/FindMedia/Settings/FindMediaSettings';
-import LibraryMapping from '../components/modules/PMS/LibraryMapping/LibraryMapping';
+import SettingsLibraryMapping from '../components/modules/Main/Settings/SettingsLibraryMapping';
+import SettingsGeneral from '../components/modules/Main/Settings/SettingsGeneral';
+import SettingsPMS from '../components/modules/Main/Settings/SettingsPMS';
+import SettingsExport from '../components/modules/Main/Settings/SettingsExport';
 import ViewState from '../components/modules/PMS/ViewState/ViewState';
 import About from '../components/modules/Main/About';
 import Store from '../store/index.js';
@@ -110,21 +113,39 @@ Vue.use(VueRouter)
     meta: {requiresAuth: true}
   },
   {
-    path: '/pms/libmapping',
-    name: "LibraryMapping",
-    component: LibraryMapping,
+    path: '/settings/settings',
+    name: "settingsGlobal",
+    component: SettingsGlobal,
+    meta: {requiresAuth: true}
+  },
+  {
+    path: '/settings/libmapping',
+    name: "SettingsLibraryMapping",
+    component: SettingsLibraryMapping,
+    meta: {requiresAuth: true}
+  },
+  {
+    path: '/settings/export',
+    name: "SettingsExport",
+    component: SettingsExport,
+    meta: {requiresAuth: true}
+  },
+  {
+    path: '/settings/settingsgeneral',
+    name: "SettingsGeneral",
+    component: SettingsGeneral,
+    meta: {requiresAuth: true}
+  },
+  {
+    path: '/settings/settingspms',
+    name: "SettingsPMS",
+    component: SettingsPMS,
     meta: {requiresAuth: true}
   },
   {
     path: '/language',
     name: "language",
     component: Language,
-    meta: {requiresAuth: true}
-  },
-  {
-    path: '/settings',
-    name: "settings",
-    component: GlobalSettings,
     meta: {requiresAuth: true}
   },
   {
